+++ crates/ffi/src/lib.rs
@@ -1,100 +1,42 @@
-mod actions;
-mod apps;
+pub(crate) mod actions;
+pub(crate) mod adapter;
+pub(crate) mod apps;
 pub(crate) mod convert;
 pub mod error;
-mod input;
-mod screenshot;
-mod surfaces;
-mod tree;
+pub(crate) mod input;
+pub(crate) mod screenshot;
+pub(crate) mod surfaces;
+pub(crate) mod tree;
 pub mod types;
-mod windows;
+pub(crate) mod windows;
 
+pub use adapter::AdAdapter;
 pub use error::AdResult;
-pub use types::*;
-
-use agent_desktop_core::adapter::PlatformAdapter;
-
-pub struct AdAdapter {
-    inner: Box<dyn PlatformAdapter>,
-}
-
-fn build_adapter() -> Box<dyn PlatformAdapter> {
-    #[cfg(target_os = "macos")]
-    {
-        Box::new(agent_desktop_macos::MacOSAdapter::new())
-    }
-
-    #[cfg(target_os = "windows")]
-    {
-        Box::new(agent_desktop_windows::WindowsAdapter::new())
-    }
-
-    #[cfg(target_os = "linux")]
-    {
-        Box::new(agent_desktop_linux::LinuxAdapter::new())
-    }
-
-    #[cfg(not(any(target_os = "macos", target_os = "windows", target_os = "linux")))]
-    compile_error!("Unsupported platform")
-}
-
-#[no_mangle]
-pub extern "C" fn ad_adapter_create() -> *mut AdAdapter {
-    let adapter = AdAdapter {
-        inner: build_adapter(),
-    };
-    Box::into_raw(Box::new(adapter))
-}
-
-/// # Safety
-///
-/// `adapter` must be a pointer returned by `ad_adapter_create`, or null.
-/// After this call the pointer is invalid and must not be used.
-#[no_mangle]
-pub unsafe extern "C" fn ad_adapter_destroy(adapter: *mut AdAdapter) {
-    if !adapter.is_null() {
-        drop(Box::from_raw(adapter));
-    }
-}
-
-/// # Safety
-///
-/// `adapter` must be a non-null pointer returned by `ad_adapter_create` that
-/// has not yet been destroyed.
-#[no_mangle]
-pub unsafe extern "C" fn ad_check_permissions(adapter: *const AdAdapter) -> AdResult {
-    let adapter = &*adapter;
-    match adapter.inner.check_permissions() {
-        agent_desktop_core::adapter::PermissionStatus::Granted => {
-            error::clear_last_error();
-            AdResult::Ok
-        }
-        agent_desktop_core::adapter::PermissionStatus::Denied { suggestion } => {
-            error::set_last_error(
-                &agent_desktop_core::error::AdapterError::new(
-                    agent_desktop_core::error::ErrorCode::PermDenied,
-                    "Accessibility permission not granted",
-                )
-                .with_suggestion(suggestion),
-            );
-            AdResult::ErrPermDenied
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_adapter_create_destroy() {
-        let ptr = ad_adapter_create();
-        assert!(!ptr.is_null());
-        unsafe { ad_adapter_destroy(ptr) };
-    }
-
-    #[test]
-    fn test_destroy_null_is_noop() {
-        unsafe { ad_adapter_destroy(std::ptr::null_mut()) };
-    }
-}
+pub use types::action::AdAction;
+pub use types::action_kind::AdActionKind;
+pub use types::action_result::AdActionResult;
+pub use types::app_info::AdAppInfo;
+pub use types::direction::AdDirection;
+pub use types::drag_params::AdDragParams;
+pub use types::element_state::AdElementState;
+pub use types::image_buffer::AdImageBuffer;
+pub use types::image_format::AdImageFormat;
+pub use types::key_combo::AdKeyCombo;
+pub use types::modifier::AdModifier;
+pub use types::mouse_button::AdMouseButton;
+pub use types::mouse_event::AdMouseEvent;
+pub use types::mouse_event_kind::AdMouseEventKind;
+pub use types::native_handle::AdNativeHandle;
+pub use types::node::AdNode;
+pub use types::node_tree::AdNodeTree;
+pub use types::point::AdPoint;
+pub use types::rect::AdRect;
+pub use types::ref_entry::AdRefEntry;
+pub use types::screenshot_kind::AdScreenshotKind;
+pub use types::screenshot_target::AdScreenshotTarget;
+pub use types::scroll_params::AdScrollParams;
+pub use types::surface_info::AdSurfaceInfo;
+pub use types::tree_options::AdTreeOptions;
+pub use types::window_info::AdWindowInfo;
+pub use types::window_op::AdWindowOp;
+pub use types::window_op_kind::AdWindowOpKind;
